pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gmoo-sdk"
version = "2.0.0"
description = "A Python SDK for global multi-objective optimization"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Matt Freeman", email = "matt.freeman@example.com"},
    {name = "Jason Hopkins", email = "jason.hopkins@example.com"},
]
keywords = ["optimization", "multi-objective", "engineering", "inverse-modeling"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Mathematics",
]

dependencies = [
    "numpy>=1.20.0",
    "python-dotenv>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=22.0",
    "isort>=5.0",
    "mypy>=0.910",
    "flake8>=4.0",
]
examples = [
    "matplotlib>=3.0",  # If examples need plotting
    "pandas>=1.3.0",    # For data handling in examples
]
all = [
    "gmoo-sdk[dev]",
    "gmoo-sdk[examples]",
]

[project.urls]
Homepage = "https://github.com/yourusername/python-wrapper-sidefork"
Documentation = "https://github.com/yourusername/python-wrapper-sidefork#readme"
Repository = "https://github.com/yourusername/python-wrapper-sidefork"
Issues = "https://github.com/yourusername/python-wrapper-sidefork/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
gmoo_sdk = ["py.typed"]  # For type hints

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --cov=gmoo_sdk --cov-report=term-missing"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["gmoo_sdk"]
src_paths = ["src", "tests", "examples"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true